Reorganized Files and updated paths
- Updated paths to fix deployment errors and to reflect the changes for moving the Icons, Logo and Navbar components
- Fixed another lint error that was missed last commit
